Refactors Card to a server component
