+++ app/(gcforms)/[locale]/(form administration)/forms/[[...path]]/components/client/MenuDropdownButton.tsx
@@ -8,7 +8,7 @@ import { useTranslation } from "react-i18next";
 import copy from "copy-to-clipboard";
 import { getForm } from "../../actions";
 import { getDate, slugify } from "@lib/client/clientHelpers";
-import { CardProps } from "./Card";
+import { CardProps } from "../server/Card";
 import { ConfirmDelete } from "@clientComponents/form-builder/app/ConfirmDelete";
 import { useCallback, useState } from "react";
 import { useRefresh } from "@lib/hooks";
